+++ crates/holochain/src/core/queue_consumer/integrate_dht_ops_consumer.rs
@@ -3,16 +3,18 @@
 use super::*;
 use crate::conductor::manager::TaskManagerClient;
 use crate::core::workflow::integrate_dht_ops_workflow::integrate_dht_ops_workflow;
+use holochain_state::dht_store::DhtStore;
 use std::sync::Arc;
 
 /// Spawn the QueueConsumer for DhtOpIntegration workflow
 #[cfg_attr(
     feature = "instrument",
-    tracing::instrument(skip(env, trigger_receipt, tm, network, conductor))
+    tracing::instrument(skip(env, dht_store, trigger_receipt, tm, network, conductor))
 )]
 pub fn spawn_integrate_dht_ops_consumer(
     dna_hash: Arc<DnaHash>,
     env: DbWrite<DbKindDht>,
+    dht_store: DhtStore,
     tm: TaskManagerClient,
     trigger_receipt: TriggerSender,
     network: DynHolochainP2pDna,
@@ -28,6 +30,7 @@ pub fn spawn_integrate_dht_ops_consumer(
         move || {
             integrate_dht_ops_workflow(
                 env.clone(),
+                dht_store.clone(),
                 trigger_receipt.clone(),
                 network.clone(),
                 conductor.clone(),